\name{findB.design}
\alias{findB.design}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Calculates B-value stopping boundary}
\description{
Internal function called by MCbound. Note same function also calculates stopping boundary for 
truncated sequential probability ratio test}
\author{M.P.Fay}
\keyword{internal}% at least one, from doc/KEYWORDS
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
