\name{MCbound.type}
\alias{MCbound.type}
\alias{MCbound.fixed}
\alias{MCbound.tsprt}
\alias{MCbound.Bvalue}
\alias{MCbound.BC}
\title{Create Monte Carlo Stopping Boundary: functions by type}
\description{
These functions are called by \code{\link{MCbound}}, see that help for details.
}
\author{Michael P. Fay}
\keyword{internal}% at least one, from doc/KEYWORDS
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
