\name{MCtest.default}
\alias{MCtest.default}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Internal function to calculation MC test}
\description{
Called by \code{\link{MCtest}}
}
\usage{
MCtest.default(x, statistic, resample, bound, extreme = "geq", seed = 1234325)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
See \code{\link{MCtest}}
}
\author{M.P. Fay}
\seealso{  \code{\link{MCtest}}}
\keyword{internal}% at least one, from doc/KEYWORDS
%\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
